+++ components/hal/esp32c2/include/hal/sha_ll.h
@@ -138,7 +138,7 @@ static inline void sha_ll_write_digest(esp_sha_type sha_type, void *digest_state
     uint32_t *digest_state_words = (uint32_t *)digest_state;
     uint32_t *reg_addr_buf = (uint32_t *)(SHA_H_BASE);
 
-    for (int i = 0; i < digest_word_len; i++) {
+    for (int i = 0; i < (int)digest_word_len; i++) {
         REG_WRITE(&reg_addr_buf[i], digest_state_words[i]);
     }
 }
